edit tool changes
